# Ensure retain(...) will always throw once it did once.