--- a/common/src/main/java/io/netty/util/AbstractReferenceCounted.java
+++ b/common/src/main/java/io/netty/util/AbstractReferenceCounted.java
@@ -18,6 +18,7 @@ package io.netty.util;
 import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
 
 import static io.netty.util.internal.ObjectUtil.checkPositive;
+import static java.lang.Math.max;
 
 /**
  * Abstract base class for classes wants to implement {@link ReferenceCounted}.
@@ -27,11 +28,13 @@ public abstract class AbstractReferenceCounted implements ReferenceCounted {
     private static final AtomicIntegerFieldUpdater<AbstractReferenceCounted> refCntUpdater =
             AtomicIntegerFieldUpdater.newUpdater(AbstractReferenceCounted.class, "refCnt");
 
+    @SuppressWarnings("FieldMayBeFinal")
     private volatile int refCnt = 1;
 
     @Override
     public final int refCnt() {
-        return refCnt;
+        // Never return anything smaller then 0 to give the user a consistent few all the time.
+        return max(0, refCntUpdater.get(this));
     }
 
     /**
@@ -52,13 +55,24 @@ public abstract class AbstractReferenceCounted implements ReferenceCounted {
     }
 
     private ReferenceCounted retain0(int increment) {
-        int oldRef = refCntUpdater.getAndAdd(this, increment);
-        if (oldRef <= 0 || oldRef + increment < oldRef) {
-            // Ensure we don't resurrect (which means the refCnt was 0) and also that we encountered an overflow.
-            refCntUpdater.getAndAdd(this, -increment);
-            throw new IllegalReferenceCountException(oldRef, increment);
+        int oldRef = refCntUpdater.get(this);
+        if (oldRef <= 0 || oldRef + increment < 0) {
+            // Either already released or we overflow.
+            throw newReferenceCountException(oldRef, increment);
         }
-        return this;
+        int ref = refCntUpdater.addAndGet(this, increment);
+        if (ref > increment) {
+            // Most likely code-path to hit.
+            return this;
+        }
+        if (ref <= 0) {
+            // Overflow happened, trying to recover
+            refCntUpdater.addAndGet(this, -increment);
+            throw newReferenceCountException(ref, increment);
+        }
+        // Set the reference count back to 0 which signals this reference was deallocated already.
+        refCntUpdater.set(this, 0);
+        throw newReferenceCountException(ref, increment);
     }
 
     @Override
@@ -79,14 +93,22 @@ public abstract class AbstractReferenceCounted implements ReferenceCounted {
     private boolean release0(int decrement) {
         int oldRef = refCntUpdater.getAndAdd(this, -decrement);
         if (oldRef == decrement) {
+            // Most likely code-path to hit.
             deallocate();
             return true;
-        } else if (oldRef < decrement || oldRef - decrement > oldRef) {
+        }
+
+        if (oldRef < decrement || oldRef - decrement > oldRef) {
             // Ensure we don't over-release, and avoid underflow.
-            refCntUpdater.getAndAdd(this, decrement);
-            throw new IllegalReferenceCountException(oldRef, -decrement);
+            // Also set the reference count back to 0 as this signals we already completely released it.
+            refCntUpdater.set(this, 0);
+            throw newReferenceCountException(oldRef, -decrement);
         }
         return false;
+    }
+
+    private static IllegalReferenceCountException newReferenceCountException(int oldCnt, int cnt) {
+        return new IllegalReferenceCountException(max(0, oldCnt), cnt);
     }
 
     /**