@normanmaurer 2018-11-19
@carl-mastrangelo @merlimat @njhill let me know what you think... it costs a bit of performance but I think it is still not so bad.