@njhill 2018-11-19
@normanmaurer I can see some potential issues with this approach:
- I think there is still a race condition whereby subsequent calls could succeed following `release()` returning true:
   - refcount starts at 1. T1 calls `retain()` and passes the initial check on line 89
   - T2 calls `release()` which returns true, refcount is now 0
   - T2 calls `release()` again, refcount is decremented to -1 (does not yet reach line 104)
   - T1 performs `addAndGet()` which increments refcount to 0
   - T2 sets refcount to 0 on line 104 (it's already 0) and throws `IRCE`
   - T1 calls `addAndGet()` which "reverts" refcount to 1, throws `IRCE`
   - Subsequent `retain()`/`release()` will succeed; buffer will be deallocated a second time
- The semantics change w.r.t. invalid releases - previously `release(2)` on a refcount of 1 would throw and leave the state unchanged. With this change, it would still throw but result in the refcount ending up as 0 _without_ `deallocate()` being called, i.e. the buffer ends up in an un-releaseable state and never deallocated

The approach referenced in https://github.com/netty/netty/issues/8563#issuecomment-439987189 avoids these problems, but likely has a slightly bigger perf hit in high contention case. I can do the benchmarks but didn't want to waste time if it was a non-starter.