@njhill 2018-11-19
Similar example, maybe clearer:
   - Refcount starts at 1. T1 calls `retain()` and passes initial check
   - T2 calls `release()` which returns true, refcount now 0
   - T1 performs `addAndGet()` which increments refcount to 1
   - T2 calls `release()` again which returns true again