@normanmaurer 2018-11-19
@njhill true... benchmark would be nice