@normanmaurer 2018-11-20
Closing this one and waiting for @njhill to open a PR based on 
https://github.com/netty/netty/commit/0c0e092e8588d5c6729080b5ae7345e041470616